Refactor audio experiment into its own page
The idea of audio-based tracking wasn't working so well because
it seems there was auto-gain-control that is not possible
to turn off (at least with the API I have).

So now the focus of the tracker work will be on the orientation
and video streams, which seem to be working well now.