Fix tracker demo to use device orientation
Seems that the matrix calculated by the code snippet in the
standards document is not correct (or somehow is not consistent
with what Slicer/VTK considers to be a rotation matrix).
Rotation around the axis perpendiular to the screen does not appear
correctly, while rotations around the two axes perpendicular
to that do appear correctly.

 http://w3c.github.io/deviceorientation/spec-source-orientation.html#deviceorientation