Use slice instead of offset to construct array
Avoids a RangeError in the case where the number of bytes
in the nrrd header is odd (because the underlying
buffer for the overall Int16Array cannot be odd even if
the view itself is odd).  This makes some sense as a word
alignment issue for a TypedArray that is a view of the buffer.

This probably results in an extra copy of the data but it seems
to have no timing impact even for a full CT volume.