Add fiber bundle support and fix a-frame vr mode
Fiber bundles are now exported like models but do
not yet have custom colors.

The exported data now uses an extension to the glTF
standard for common material types instead of the
shader code copied from the glTF samples.
This allows VR mode in A-Frame to properly render
the exported objects.